@@ -1,11 +1,10 @@
 {
   "name": "@sensenet/react-typescript-boilerplate",
   "version": "1.0.2",
-  "main": "dist/index.js",
-  "files": [
-    "dist",
-    "src"
-  ],
+  "private": true,
+  "author": "Sense/Net <snteam@sensenet.com> (http://sensenet.com/)",
+  "license": "GPL-2.0",
+  "homepage": "https://sensenet.com",
   "keywords": [
     "sensenet",
     "cps",
@@ -15,51 +14,36 @@
     "typescript"
   ],
   "scripts": {
-    "fix:prettier": "prettier \"{,!(dist|temp|bundle)/**/}*.{ts,tsx}\" --write",
-    "build": "rimraf bundle && webpack",
-    "start": "webpack-dev-server --progress",
-    "test": "jest --coverage",
-    "lint": "eslint . --ext .tsx,.ts --cache"
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "eject": "react-scripts eject"
   },
-  "private": true,
-  "publishConfig": {
-    "access": "private"
-  },
-  "author": "Sense/Net <snteam@sensenet.com> (http://sensenet.com/)",
-  "license": "GPL-2.0",
-  "homepage": "https://sensenet.com",
   "devDependencies": {
-    "@types/enzyme": "^3.10.3",
     "@types/jest": "^24.0.23",
-    "@types/react": "^16.9.11",
+    "@types/node": "^12.12.14",
+    "@types/react": "^16.9.13",
     "@types/react-dom": "^16.9.4",
-    "awesome-typescript-loader": "^5.2.1",
-    "css-loader": "^3.2.0",
-    "enzyme": "^3.10.0",
-    "enzyme-adapter-react-16": "^1.15.1",
-    "enzyme-to-json": "^3.4.3",
-    "file-loader": "^4.2.0",
-    "html-webpack-plugin": "^3.2.0",
-    "jest": "^24.9.0",
-    "rimraf": "^3.0.0",
-    "source-map-loader": "^0.2.4",
-    "style-loader": "^1.0.0",
-    "ts-config-webpack-plugin": "^1.4.0",
-    "ts-jest": "^24.1.0",
-    "typescript": "^3.7.2",
-    "url-loader": "^2.2.0",
-    "webpack": "^4.41.2",
-    "webpack-bundle-analyzer": "^3.6.0",
-    "webpack-cli": "^3.3.10",
-    "webpack-dev-server": "^3.9.0"
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
+    "react-scripts": "^3.2.0",
+    "typescript": "^3.7.2"
   },
   "dependencies": {
     "@material-ui/core": "^4.6.1",
     "@sensenet/client-core": "^2.2.4",
     "@sensenet/default-content-types": "^2.0.3",
-    "@sensenet/hooks-react": "^1.0.3",
-    "react": "^16.11.0",
-    "react-dom": "^16.11.0",
-    "tslib": "^1.10.0"
+    "@sensenet/hooks-react": "^1.0.3"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
