@@ -1,4 +1,3 @@
-import Chance = require('chance')
 import { resources } from '../../../src/assets/resources'
 import {
   contextMenuItems,
@@ -11,6 +10,7 @@ import {
   selectPathInListPicker,
   uploadNewFileAndOpenContextMenuItem,
 } from '../../support/documents'
+import Chance = require('chance')
 
 context('The documents page', () => {
   let currentUser = { email: '', password: '' }
