@@ -200,11 +200,11 @@ class DocumentLibrary extends React.Component<
   }
 
   public componentDidMount() {
-    addEventListener('scroll', this.handleScroll)
+    window.addEventListener('scroll', this.handleScroll)
   }
 
   public componentWillUnmount() {
-    removeEventListener('scroll', this.handleScroll)
+    window.removeEventListener('scroll', this.handleScroll)
   }
 
   private updateLoading = debounce((value: boolean) => {
