@@ -1,5 +1,5 @@
 module.exports = {
-  coveragePathIgnorePatterns: ['index.ts'],
+  coveragePathIgnorePatterns: ['index.ts', 'examples'],
   globals: {
     'ts-jest': {
       isolatedModules: true, // comment out this and uncomment the line below to check for typescript errors
