@@ -72,7 +72,7 @@ export class AdvancedSearch<T extends GenericContent = GenericContent> extends C
     const query = new Query(q => {
       const filteredQueries = fieldQueryArray.filter(f => f.toString().length > 0)
 
-      filteredQueries.map((fieldQuery, currentIndex) => {
+      filteredQueries.forEach((fieldQuery, currentIndex) => {
         const { queryRef } = q
         new QueryExpression(queryRef).query(fieldQuery)
         if (currentIndex < filteredQueries.length - 1) {
